requirements.txt
python-telegram-bot==21.8  # Работа с Telegram Bot API
firebase-admin==6.6.0      # Работа с Firebase (Real-time Database)
python-dotenv==1.0.1      # Загрузка переменных окружения из .env
